cmd/roi: 패키지만의 HandlerFunc 타입 구현
기존 http.HandlerFunc로는 효율적인 코드의 재사용이 어려워
패키지 내의 HandlerFunc를 따로 구현하였다.

기존에 구현되어있는 핸들러들을 아직 패키지의 HandlerFunc로
변경하지는 않았다.

이는 따라오는 일련의 커밋에서 수정될 예정이다.
